+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.12] - 2022-11-08
+
+### Changed
+
+- Enforce isScrollEnabled to be true as default
+- Fixed checkout height issue
+
 ## [0.1.11] - 2022-11-08
 
 ### Added